+++ libft/sources/ft_str/ft_substr.c
@@ -3,22 +3,22 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tstrassb <tstrassb@student.42wolfsburg.    +#+  +:+       +#+        */
+/*   By: tstrassb <tstrassb@student.42>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/15 14:40:26 by tstrassb          #+#    #+#             */
-/*   Updated: 2022/12/28 09:12:39 by tstrassb         ###   ########.fr       */
+/*   Updated: 2023/03/05 09:44:30 by tstrassb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char			*str;
 
 	if (!s)
 		return (NULL);
-	if (start > (unsigned int)ft_strlen(s))
+	if (start > (int)ft_strlen(s))
 		return (ft_strdup(""));
 	if (ft_strlen(s + start) < len)
 		len = ft_strlen(s + start);
